# docs/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GymCRM API",
    "description": "Multi-tenant gym SaaS. All business endpoints require Bearer JWT. gym_id is extracted from JWT — every query is automatically scoped to the authenticated gym.",
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter: Bearer <access_token>"
    }
  },
...
  "paths": {
...
    "/api/v1/auth/register": {
      "post": { "tags": ["auth"], "summary": "Register a new gym",
        "consumes": ["application/json"], "produces": ["application/json"],
        "parameters": [{ "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/RegisterGymRequest" } }],
        "responses": { "201": { "description": "Created", "schema": { "$ref": "#/definitions/RegisterResponse" } }, "409": { "description": "Phone already registered" }, "422": { "description": "Validation error" } } }
    },
    "/api/v1/auth/login": {
      "post": { "tags": ["auth"], "summary": "Login",
        "consumes": ["application/json"], "produces": ["application/json"],
        "parameters": [{ "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/LoginRequest" } }],
        "responses": { "200": { "description": "OK", "schema": { "$ref": "#/definitions/AuthResponse" } }, "401": { "description": "Invalid credentials" } } }
    },
    "/api/v1/auth/refresh": {
      "post": { "tags": ["auth"], "summary": "Refresh access token",
        "consumes": ["application/json"], "produces": ["application/json"],
        "parameters": [{ "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/RefreshRequest" } }],
        "responses": { "200": { "description": "OK", "schema": { "$ref": "#/definitions/AuthResponse" } }, "401": { "description": "Invalid token" } } }
    },
    "/api/v1/auth/me": {
      "get": { "tags": ["auth"], "summary": "Get current user and gym",
        "security": [{ "BearerAuth": [] }], "produces": ["application/json"],
        "responses": { "200": { "description": "OK", "schema": { "$ref": "#/definitions/MeResponse" } }, "401": { "description": "Not authenticated" } } }
    },
    "/api/v1/auth/logout": {
      "post": { "tags": ["auth"], "summary": "Logout",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Logged out" } } }
    },
...
  },
  "definitions": {
    "RegisterGymRequest": {
      "type": "object", "required": ["gym_name", "owner_name", "phone", "password", "city", "state"],
      "properties": {
        "gym_name": { "type": "string", "example": "Singh Fitness" },
        "owner_name": { "type": "string", "example": "Gurpreet Singh" },
        "phone": { "type": "string", "example": "9876543210" },
        "password": { "type": "string", "example": "secure123" },
        "city": { "type": "string", "example": "Ludhiana" },
        "state": { "type": "string", "example": "Punjab" },
        "address": { "type": "string" }, "email": { "type": "string" }
      }
    },
    "LoginRequest": {
      "type": "object", "required": ["phone", "password"],
      "properties": { "phone": { "type": "string", "example": "9876543210" }, "password": { "type": "string", "example": "secure123" } }
    },
    "RefreshRequest": {
      "type": "object", "required": ["refresh_token"],
      "properties": { "refresh_token": { "type": "string" } }
    },
    "AuthResponse": {
      "type": "object",
      "properties": {
        "success": { "type": "boolean" },
        "data": { "type": "object", "properties": {
          "access_token": { "type": "string" }, "refresh_token": { "type": "string" },
          "token_type": { "type": "string", "example": "Bearer" }, "expires_in": { "type": "integer", "example": 900 },
          "user": { "$ref": "#/definitions/UserDTO" }
        }}
      }
    },
    "RegisterResponse": {
      "type": "object",
      "properties": {
        "success": { "type": "boolean" },
        "data": { "type": "object", "properties": {
          "gym": { "$ref": "#/definitions/GymDTO" }, "user": { "$ref": "#/definitions/UserDTO" },
          "access_token": { "type": "string" }, "refresh_token": { "type": "string" },
          "token_type": { "type": "string" }, "expires_in": { "type": "integer" }
        }}
      }
    },
    "MeResponse": {
      "type": "object",
      "properties": {
        "success": { "type": "boolean" },
        "data": { "type": "object", "properties": {
          "user": { "$ref": "#/definitions/UserDTO" }, "gym": { "$ref": "#/definitions/GymDTO" }
        }}
      }
    },
    "UserDTO": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" }, "gym_id": { "type": "integer" },
        "name": { "type": "string" }, "phone": { "type": "string" },
        "email": { "type": "string" }, "role": { "type": "string", "enum": ["owner", "staff"] }
      }
    },
    "GymDTO": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" }, "name": { "type": "string" },
        "city": { "type": "string" }, "state": { "type": "string" }, "status": { "type": "string" }
      }
    },
...
  }
}